pyproject.toml
[tool.poetry]
name = "PyAssistant"
version = "0.1.0"
description = "User interface for interacting with pretrained, locally stored AI models."
authors = ["Brent Bradford <brent.c.bradford@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
diffusers = {extras = ["torch"], version = "^0.27.2"}
transformers = {extras = ["torch"], version = "^4.40.1"}
torchvision = "^0.18.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
accelerate = "^0.30.0"
safetensors = "^0.4.3"
sentencepiece = "^0.2.0"
bitsandbytes = "^0.43.1"
beautifulsoup4 = "^4.12.3"
ftfy = "^6.2.0"

[tool.poetry.scripts]
pyassistant = "pyassistant.app:start"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
